Replace previous snapshot tree with a new snapshot tree data structure
this tree is capable of representing multiple parents more easily and is easier to maintain than the previous approach.
